Reduce random stake amount (#134)
* Reduce random stake amount

* Fix type

* Fix type

* Add comment

* Organize

* Update MockContract submit type int->float

* Assert stake amount

+++ pdr_backend/models/predictoor_contract.py
@@ -248,7 +248,7 @@ class PredictoorContract(BaseContract):  # pylint: disable=too-many-public-metho
     def submit_prediction(
         self,
         predicted_value: bool,
-        stake_amount: int,
+        stake_amount: float,
         prediction_ts: int,
         wait_for_receipt=True,
     ):
